Layout adjustments to A320 limitations page
The markup for this page was shifted around quite a bit, which might make the diff (the different between the old and new) a bit hard to read. Basically you now have all three `.col`  divs inside a single `.container` div, which is critical for the grid to function properly. Now the floats are properly contained and aligned.

I’ve also set all tables in `.tables` divs to always expand to 100%. This will also help with alignment issues. if you want to expand or contract them, make the column narrower or wider. You could make the third one less than 100% if you wanted and it would still sit below the other two.